+++ src/Autocompletion/CompletionItem.php
@@ -28,6 +28,11 @@ final class CompletionItem implements JsonSerializable, ArrayAccess
     private $kind;
 
     /**
+     * @var int[]|null Zero, one, or more items from {@see CompletionItemTag}
+     */
+    private $tags;
+
+    /**
      * @var string|null
      */
     private $insertText;
@@ -83,6 +88,7 @@ final class CompletionItem implements JsonSerializable, ArrayAccess
      * @param bool          $deprecated
      * @param string|null   $detail
      * @param string|null   $sortText
+     * @param int[]|null    $tags
      */
     public function __construct(
         string $filterText,
@@ -94,7 +100,8 @@ final class CompletionItem implements JsonSerializable, ArrayAccess
         array $additionalTextEdits = [],
         bool $deprecated = false,
         ?string $detail = null,
-        ?string $sortText = null
+        ?string $sortText = null,
+        ?array $tags = null
     ) {
         $this->filterText = $filterText;
         $this->kind = $kind;
@@ -106,6 +113,12 @@ final class CompletionItem implements JsonSerializable, ArrayAccess
         $this->deprecated = $deprecated;
         $this->detail = $detail;
         $this->sortText = $sortText;
+
+        if ($tags === null && $deprecated) {
+            $this->tags = [CompletionItemTag::DEPRECATED];
+        } else {
+            $this->tags = $tags;
+        }
 
         if ($insertText === null && $textEdit === null) {
             throw new LogicException('Either an insertText or a textEdit must be provided');
@@ -126,6 +139,14 @@ final class CompletionItem implements JsonSerializable, ArrayAccess
     public function getKind(): ?int
     {
         return $this->kind;
+    }
+
+    /**
+     * @return int[]|null
+     */
+    public function getTags(): ?array
+    {
+        return $this->tags;
     }
 
     /**
@@ -208,6 +229,7 @@ final class CompletionItem implements JsonSerializable, ArrayAccess
         return [
             'filterText'          => $this->getFilterText(),
             'kind'                => $this->getKind(),
+            'tags'                => $this->getTags(),
             'insertText'          => $this->getInsertText(),
             'insertTextFormat'    => $this->getInsertTextFormat(),
             'sortText'            => $this->getSortText(),