+++ src/Autocompletion/CompletionItemTag.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Serenata\Autocompletion;
+
+/**
+ * Enumeration of suggestion tag.
+ */
+final class CompletionItemTag
+{
+    /**
+     * @var int
+     */
+    public const DEPRECATED = 1;
+}